feat: support embed of javascript files
- Big feature: dynamic loading and listing of JS files on every page

As an example, take a look at server-time-to-local.js. It formats the time into a specific format that the user can select on the /account or /update-account routes (settings page).

You can also add any number of JS files to /assets/js/ folder and they will all be loaded on every page (which at the same time is a limitation of this particular method).

Every JS file is embedded into a virtual file system and then served from there to the client in a separate request.

----

- Updated the Dockerfile to use more modern Go.
- Added pre-go-generated files, too.